refactor: Product 패키지 분리
