+++ src/main/java/kitchenpos/product/repository/JdbcTemplateProductDao.java
@@ -1,9 +1,10 @@
-package kitchenpos.dao;
+package kitchenpos.product.repository;
 
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.repository.ProductDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
