+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,10 +1,10 @@
-package kitchenpos.ui;
+package kitchenpos.product.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.application.ProductService;
-import kitchenpos.application.dto.request.ProductCreateRequest;
-import kitchenpos.application.dto.response.ProductResponse;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.ui.request.ProductCreateRequest;
+import kitchenpos.product.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
