+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -4,8 +4,9 @@ import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderLineItemDao;
 import kitchenpos.dao.OrderTableDao;
-import kitchenpos.dao.ProductDao;
+import kitchenpos.product.repository.ProductDao;
 import kitchenpos.dao.TableGroupDao;
+import kitchenpos.product.application.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
