+++ src/test/java/kitchenpos/product/doamin/ProductTest.java
@@ -1,10 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.product.doamin;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.math.BigDecimal;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
